extracted PasswordField component
